Moved test survey under surveys to be autogenerated
