+++ app/surveys/large_components/pages/page6.py
@@ -3,7 +3,7 @@ from app.model.section import Section
 from app.model.question import Question
 from app.model.response import Response
 
-from app.test_survey.test_survey import test_questionnaire
+from app.surveys.large_test import test_questionnaire
 
 page6 = Block()
 page6.id = 'page-6'
